# apps/lumistripe-web/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioCalibrationFinishRequest": {
        "properties": {
          "apply": {
            "title": "Apply",
            "type": "boolean"
          },
          "noise_floor": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Noise Floor"
          },
          "target_level": {
            "anyOf": [
              {
                "maximum": 0.8,
                "minimum": 0.1,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Level"
          }
        },
        "required": [
          "apply"
        ],
        "title": "AudioCalibrationFinishRequest",
        "type": "object"
      },
...
      "AudioCalibrationSessionResponse": {
        "properties": {
          "elapsed_seconds": {
            "default": 0.0,
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "remaining_seconds": {
            "default": 0.0,
            "title": "Remaining Seconds",
            "type": "number"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AudioCalibrationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status"
        ],
        "title": "AudioCalibrationSessionResponse",
        "type": "object"
      },
      "AudioCalibrationStartRequest": {
        "properties": {
          "device": {
            "minLength": 1,
            "title": "Device",
            "type": "string"
          },
          "duration_seconds": {
            "default": 8.0,
            "maximum": 30.0,
            "minimum": 3.0,
            "title": "Duration Seconds",
            "type": "number"
          }
        },
        "required": [
          "device"
        ],
        "title": "AudioCalibrationStartRequest",
        "type": "object"
      },
...
      "AudioDeviceRequest": {
        "properties": {
          "device": {
            "minLength": 1,
            "title": "Device",
            "type": "string"
          }
        },
        "required": [
          "device"
        ],
        "title": "AudioDeviceRequest",
        "type": "object"
      },
...
      "AudioOutputMuteRequest": {
        "properties": {
          "muted": {
            "title": "Muted",
            "type": "boolean"
          },
          "selector": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Selector",
            "type": "string"
          }
        },
        "required": [
          "selector",
          "muted"
        ],
        "title": "AudioOutputMuteRequest",
        "type": "object"
      },
      "AudioOutputSelectionRequest": {
        "properties": {
          "selector": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Selector",
            "type": "string"
          }
        },
        "required": [
          "selector"
        ],
        "title": "AudioOutputSelectionRequest",
        "type": "object"
      },
      "AudioOutputVolumeRequest": {
        "properties": {
          "selector": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Selector",
            "type": "string"
          },
          "volume": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "selector",
          "volume"
        ],
        "title": "AudioOutputVolumeRequest",
        "type": "object"
      },
...
      "BluetoothAliasRequest": {
        "properties": {
          "alias": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Alias",
            "type": "string"
          }
        },
        "required": [
          "alias"
        ],
        "title": "BluetoothAliasRequest",
        "type": "object"
      },
...
      "BluetoothDeviceRequest": {
        "properties": {
          "address": {
            "maxLength": 17,
            "minLength": 17,
            "title": "Address",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "input",
                  "output"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "required": [
          "address"
        ],
        "title": "BluetoothDeviceRequest",
        "type": "object"
      },
      "BluetoothPowerRequest": {
        "properties": {
          "powered": {
            "title": "Powered",
            "type": "boolean"
          }
        },
        "required": [
          "powered"
        ],
        "title": "BluetoothPowerRequest",
        "type": "object"
      },
      "BluetoothStatusResponse": {
        "properties": {
          "adapter_alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Alias"
          },
          "available": {
            "default": false,
            "title": "Available",
            "type": "boolean"
          },
          "capabilities": {
            "$ref": "#/components/schemas/BluetoothCapabilities"
          },
          "connected_device": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BluetoothDeviceInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "connected_inputs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BluetoothDeviceInfo"
            },
            "title": "Connected Inputs",
            "type": "array"
          },
          "connected_outputs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BluetoothDeviceInfo"
            },
            "title": "Connected Outputs",
            "type": "array"
          },
          "default_sink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Sink"
          },
          "devices": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BluetoothDeviceInfo"
            },
            "title": "Devices",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Source"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "operation_state": {
            "default": "idle",
            "enum": [
              "idle",
              "running",
              "complete",
              "failed"
            ],
            "title": "Operation State",
            "type": "string"
          },
          "output_devices": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AudioOutputDeviceInfo"
            },
            "title": "Output Devices",
            "type": "array"
          },
          "output_muted": {
            "default": false,
            "title": "Output Muted",
            "type": "boolean"
          },
          "output_ready": {
            "default": false,
            "title": "Output Ready",
            "type": "boolean"
          },
          "output_volume": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Volume"
          },
          "powered": {
            "default": false,
            "title": "Powered",
            "type": "boolean"
          },
          "scanning": {
            "default": false,
            "title": "Scanning",
            "type": "boolean"
          },
          "streaming": {
            "default": false,
            "title": "Streaming",
            "type": "boolean"
          }
        },
        "title": "BluetoothStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LumiStripe",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/audio/bluetooth/alias": {
      "put": {
        "operationId": "bluetooth_alias_api_audio_bluetooth_alias_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothAliasRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Alias"
      }
    },
    "/api/audio/bluetooth/connect": {
      "post": {
        "operationId": "bluetooth_connect_api_audio_bluetooth_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Connect"
      }
    },
    "/api/audio/bluetooth/disconnect": {
      "post": {
        "operationId": "bluetooth_disconnect_api_audio_bluetooth_disconnect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Disconnect"
      }
    },
    "/api/audio/bluetooth/forget": {
      "post": {
        "operationId": "bluetooth_forget_api_audio_bluetooth_forget_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Forget"
      }
    },
    "/api/audio/bluetooth/pair": {
      "post": {
        "operationId": "bluetooth_pair_api_audio_bluetooth_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Pair"
      }
    },
    "/api/audio/bluetooth/power": {
      "put": {
        "operationId": "bluetooth_power_api_audio_bluetooth_power_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BluetoothPowerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bluetooth Power"
      }
    },
...
    "/api/audio/calibration/session": {
      "post": {
        "operationId": "start_audio_calibration_api_audio_calibration_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioCalibrationStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioCalibrationSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Audio Calibration"
      }
    },
    "/api/audio/calibration/session/{session_id}": {
      "get": {
        "operationId": "audio_calibration_status_api_audio_calibration_session__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioCalibrationSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audio Calibration Status"
      }
    },
    "/api/audio/calibration/session/{session_id}/finish": {
      "post": {
        "operationId": "finish_audio_calibration_api_audio_calibration_session__session_id__finish_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioCalibrationFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish Audio Calibration"
      }
    },
    "/api/audio/device": {
      "put": {
        "operationId": "select_audio_device_api_audio_device_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Audio Device"
      }
    },
    "/api/audio/output": {
      "put": {
        "operationId": "select_audio_output_api_audio_output_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioOutputSelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Audio Output"
      }
    },
    "/api/audio/output/mute": {
      "put": {
        "operationId": "set_audio_output_mute_api_audio_output_mute_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioOutputMuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Audio Output Mute"
      }
    },
    "/api/audio/output/volume": {
      "put": {
        "operationId": "set_audio_output_volume_api_audio_output_volume_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioOutputVolumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BluetoothStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Audio Output Volume"
      }
    },
    "/api/audio/settings": {
      "get": {
        "operationId": "audio_settings_api_audio_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Audio Settings"
      },
      "put": {
        "operationId": "update_audio_settings_api_audio_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Audio Settings"
      }
    },
    "/api/audio/settings/reset": {
      "post": {
        "operationId": "reset_audio_settings_api_audio_settings_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Audio Settings"
      }
    },
    "/api/audio/source": {
      "put": {
        "operationId": "select_audio_source_api_audio_source_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Audio Source"
      }
    },
...
  }
}